move page externsion to own class
